details view modularized
